feat(0070): adding clickable pois feature
