+++ expo-maps/android/src/main/java/expo/modules/maps/googleMaps/GoogleMapsPlaces.kt
@@ -1,7 +1,9 @@
 package expo.modules.maps.googleMaps
 
 import android.content.Context
+import android.widget.Toast
 import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.PointOfInterest
 import com.google.android.libraries.places.api.*
 import com.google.android.libraries.places.api.net.*
 import expo.modules.kotlin.Promise
@@ -9,12 +11,13 @@ import expo.modules.maps.googleMaps.placesUtils.GoogleMapsPlacesSearchCompleter
 import expo.modules.maps.googleMaps.placesUtils.GoogleMapsPlacesTokenUtils
 import expo.modules.maps.googleMaps.placesUtils.GooglePlacesFetchPlace
 
-class GoogleMapsPlaces(private val context: Context, private val map: GoogleMap, private val markers: GoogleMapsMarkers) {
+class GoogleMapsPlaces(private val context: Context, map: GoogleMap, private val markers: GoogleMapsMarkers): GoogleMap.OnPoiClickListener {
 
     private val placesClient: PlacesClient
     private val tokenUtils: GoogleMapsPlacesTokenUtils
     private val placesSearchCompleter: GoogleMapsPlacesSearchCompleter
     private val placesFetcher: GooglePlacesFetchPlace
+    private var clickablePOIs: Boolean = true
 
     init {
         Places.initialize(context, "apiKey")
@@ -42,9 +45,19 @@ class GoogleMapsPlaces(private val context: Context, private val map: GoogleMap,
         return if (tmpStr.size > 1) tmpStr[1] else ""
     }
 
+    fun setClickablePOIs(clickablePOIs: Boolean) {
+        this.clickablePOIs = clickablePOIs
+    }
 
-
-
+    override fun onPoiClick(poi: PointOfInterest) {
+        if (clickablePOIs) {
+            Toast.makeText(this.context, """Clicked: ${poi.name}
+            Place ID:${poi.placeId}
+            Latitude:${poi.latLng.latitude} Longitude:${poi.latLng.longitude}""",
+                    Toast.LENGTH_SHORT
+            ).show()
+        }
+    }
 
 
 }
